// Admin/src/models/reportModel.tsx
export const reportApiDetails = {
  endpoint: '/rcm/reports',
  filterQueryParam: 'filter',
  responseDataPath: 'data',
  responseTotalPath: 'meta.total',
} as const